+++ lib/core/web/url_strategy.dart
@@ -1,0 +1,2 @@
+export 'url_strategy_stub.dart'
+  if (dart.library.html) 'url_strategy_web.dart';
