+++ lib/core/web/url_strategy_stub.dart
@@ -1,0 +1,1 @@
+void configureUrlStrategy () {}
