+++ lib/core/web/url_strategy_web.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+
+void configureUrlStrategy() {
+  usePathUrlStrategy();
+}
