+++ lib/telegram/telegram_web_app_stub.dart
@@ -1,0 +1,5 @@
+class TelegramWebApp {
+  bool get isAvailable => false;
+  static void ready() {}
+  static void sendData(Map<String, dynamic> data) {}
+}
